Support removing unused package files via DROP_UNUSED_DEBFILES
At a customer of mine when executing repository_checker with its
--validate-incoming option to ensure that the incoming directory
is in a sane state we noticed problems with different compression
defaults. With the recent switch to Debian/jessie for the build
infrastructure we noticed that *.tar.xz files are left behind:

| 12:12:59 Checking for leftover files in incoming directory /srv/repository/release/release-mr0.42/incoming/release-mr0.42:
| [...]
| 12:12:59 /srv/repository/release/release-mr0.42/incoming/release-mr0.42/ngcp-ngcpcfg_0.26.1.1+0~mr0.42.1.tar.xz
| [...]
| 12:12:59 Leftover files found. Needs investigation.

This is caused by the fact that executing generate-git-snapshot
on Debian/jessie by default generates a *.tar.xz file nowadays,
quoting dpkg-source(1) :

|  -Zcompression, --compression=compression
|
| 	  Specify the compression to use for created files
| 	  (tarballs and diffs).  Note that this option will not
| 	  cause existing tarballs to be recompressed, it only
| 	  affects new files. Supported  values are: gzip, bzip2,
| 	  lzma and xz.  The default is xz for formats 2.0 and
| 	  newer, and gzip for format 1.0. xz is only supported
| 	  since dpkg 1.15.5.

During the binary package build another tarball gets created via
dpkg-buildpackage's -sa option (to include the sources). In the
binary package build for an older release (like Debian/wheezy)
this won't default to *.tar.xz though. Therefore we end up with
an additional tarball, usually being *.tar.gz. Now we have a
*.tar.gz and a *.tar.xz next to each other and when simply
grabbing all files for usage as artifacts we end up with one
unneeded tarball which is identified as "leftover file(s) in
incoming directory".

We sadly can't just default to build with dpkg-buildpackages' -b
option because then we don't get the source files in *any* binary
package build. At least once the source files need to be present
though to end up properly in the archive. Until we've a default
and reliable way how to do that this workaround mitigates the
above mentioned problem.

NOTE: When uploading artifacts e.g. via dput (instead of simply
copying them around) then this isn't a problem at all since
*.changes refers to the files that should be uploaded (and unused
files aren't included). Therefore we don't enable this behaviour
as default but provide the user an option to enable it on demand.